+++ packages/std/src/math/uint256.rs
@@ -284,6 +284,13 @@ impl Uint256 {
         Self(self.0.saturating_mul(other.0))
     }
 
+    pub fn saturating_pow(self, exp: u32) -> Self {
+        match self.checked_pow(exp) {
+            Ok(value) => value,
+            Err(_) => Self::MAX,
+        }
+    }
+
     pub fn abs_diff(self, other: Self) -> Self {
         if self < other {
             other - self
@@ -1465,6 +1472,11 @@ mod tests {
             Uint256::MAX.saturating_mul(Uint256::from(2u32)),
             Uint256::MAX
         );
+        assert_eq!(
+            Uint256::from(4u32).saturating_pow(2u32),
+            Uint256::from(16u32)
+        );
+        assert_eq!(Uint256::MAX.saturating_pow(2u32), Uint256::MAX);
     }
 
     #[test]